# expression: implement vectorized evaluation for `builtinPeriodAddSig`