--- a/expression/builtin_time_vec.go
+++ b/expression/builtin_time_vec.go
@@ -1006,7 +1006,7 @@ func (b *builtinPeriodAddSig) vecEvalInt(input *chunk.Chunk, result *chunk.Colum
 			continue
 		}
 
-		// this is non-vectorized implementation
+		// in MySQL, if p is invalid but n is NULL, the result is NULL, so we have to check if n is NULL first.
 		if !validPeriod(i64s[i]) {
 			return errIncorrectArgs.GenWithStackByArgs("period_add")
 		}