--- a/expression/builtin_time_vec_test.go
+++ b/expression/builtin_time_vec_test.go
@@ -14,6 +14,7 @@
 package expression
 
 import (
+	"math/rand"
 	"testing"
 
 	. "github.com/pingcap/check"
@@ -23,6 +24,12 @@ import (
 	"github.com/pingcap/tidb/util/chunk"
 	"github.com/pingcap/tidb/util/mock"
 )
+
+type periodGener struct{}
+
+func (g *periodGener) gen() interface{} {
+	return int64((rand.Intn(2500)+1)*100 + rand.Intn(12) + 1)
+}
 
 var vecBuiltinTimeCases = map[string][]vecExprBenchCase{
 	ast.DateLiteral: {},
@@ -55,7 +62,7 @@ var vecBuiltinTimeCases = map[string][]vecExprBenchCase{
 	ast.MakeDate: {},
 	ast.MakeTime: {},
 	ast.PeriodAdd: {
-		{retEvalType: types.ETInt, childrenTypes: []types.EvalType{types.ETInt, types.ETInt}, geners: []dataGenerator{&rangeInt64Gener{1, 13}, &rangeInt64Gener{1, 20}}},
+		{retEvalType: types.ETInt, childrenTypes: []types.EvalType{types.ETInt, types.ETInt}, geners: []dataGenerator{new(periodGener), new(periodGener)}},
 	},
 	ast.PeriodDiff: {},
 	ast.Quarter:    {},