@SunRunAway 2019-11-05
> Add the comment of the non-vectorized implementation here.

Hi @shihongzhi , Please resolve this comment and resolve the conflicting files. Thank you.